chore(frontend/ResourceModal): cleanup
