+++ frontend/src/styling/NotesDisplayCard.module.scss
@@ -22,6 +22,7 @@
 
 .overlay {
   position: absolute;
+  border: none;
   top: 0;
   left: 0;
   width: 100vw;
